+++ schemas/builder/termStatus.js
@@ -1,0 +1,14 @@
+export default [
+  {
+    frTitle: 'normalisé',
+    deTitle: 'genormt',
+    itTitle: 'genormt',
+    value: 'normalized',
+  },
+  {
+    frTitle: 'désuet',
+    deTitle: 'veraltet',
+    itTitle: 'veraltet',
+    value: 'old',
+  },
+]
